@@ -1,20 +1,17 @@
 {
   "name": "mod-dep-mod",
-  "version": "3.0.1",
-  "description": "",
+  "version": "4.0.0",
+  "type": "module",
+  "description": "tool to find modules that depend on other modules, comes in handy when you don't know what is pulling in what",
   "bin": {
     "mod-dep-mod": "bin/index.js"
   },
   "scripts": {
-    "build": "npx babel src --out-dir lib --ignore \"src/**/test.js\"",
-    "watch": "npx babel --watch src --out-dir lib --ignore \"src/**/test.js\"",
-    "lint": "npx standard --env mocha src/**/*.js",
-    "test": "cross-env NODE_ENV=test npx mocha --timeout 4000 --require @babel/register src/**/*test.js",
-    "cover": "cross-env NODE_ENV=test npx nyc mocha --timeout 4000 ./src/**/test.js --require @babel/register",
-    "cover:html": "cross-env NODE_ENV=test npx nyc --reporter=html npm run cover "
+    "lint": "standard --env mocha src/**/*.js",
+    "test": "cross-env NODE_ENV=test npx mocha --loader=testdouble --timeout 4000 src/**/*.spec.js"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=16.13.1"
   },
   "repository": {
     "type": "git",
@@ -35,79 +32,29 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@babel/runtime": "7.16.5",
     "async": "3.2.2",
     "isomorphic-fetch": "3.0.0",
     "minimist": "1.2.5",
     "npm-package-arg": "8.1.5",
-    "ora": "5.4.1",
+    "ora": "^6.0.1",
     "pify": "5.0.0",
     "read-package-json": "4.1.1",
     "registry-url": "5.1.0",
     "semver": "7.3.5"
   },
   "devDependencies": {
-    "@babel/cli": "7.16.0",
-    "@babel/core": "7.16.5",
-    "@babel/plugin-transform-runtime": "7.16.5",
-    "@babel/polyfill": "7.12.1",
-    "@babel/preset-env": "7.16.5",
-    "@babel/register": "7.16.5",
-    "babel-eslint": "10.1.0",
-    "babel-plugin-rewire": "1.2.0",
     "chai": "4.3.4",
     "cross-env": "7.0.3",
     "mocha": "9.1.3",
     "nock": "13.1.4",
     "nodemon": "2.0.15",
-    "nyc": "15.1.0",
-    "standard": "16.0.4"
+    "standard": "16.0.4",
+    "testdouble": "^3.16.3"
   },
   "standard": {
-    "parser": "babel-eslint",
     "ignore": "lib",
     "env": {
       "mocha": true
-    }
-  },
-  "nyc": {
-    "require": [
-      "@babel/register"
-    ],
-    "include": [
-      "src/**/*.js"
-    ],
-    "exclude": [
-      "src/**/test.js"
-    ],
-    "report": [
-      "json",
-      "lcov",
-      "html",
-      "text"
-    ]
-  },
-  "babel": {
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "targets": {
-            "node": "6.0.0"
-          }
-        }
-      ]
-    ],
-    "plugins": [
-      "@babel/plugin-transform-runtime"
-    ],
-    "env": {
-      "test": {
-        "plugins": [
-          "babel-plugin-rewire",
-          "@babel/plugin-transform-runtime"
-        ]
-      }
     }
   }
 }
