@@ -1,8 +1,10 @@
 #!/usr/bin/env node
 
-var argv = require('minimist')(process.argv.slice(2))
-var walk = require('../lib/walkDeps').default
-
+import minimist from 'minimist'
+import walk from '../src/walkDeps/index.js'
+// var argv = require('minimist')(process.argv.slice(2))
+// var walk = require('../src/walkDeps').default
+const argv = minimist(process.argv.slice(2))
 if (argv._.length === 0) {
   console.log('no args given')
   console.log('example mod-dep-mod debug')
@@ -11,7 +13,7 @@ if (argv._.length === 0) {
 
 process.on('unhandledRejection', up => { throw new Error('uncaught!') })
 
-var getDependencyTree = function run(modules, fullyQualPck, urlToPck) {
+const getDependencyTree = function run (modules, fullyQualPck, urlToPck) {
   const pck = fullyQualPck || urlToPck || 'package.json'
   walk(modules, pck, process.env.NODE_ENV, () => process.exit())
     .then(() => null)
